requirements.txt
rawpy~=0.18.1
Pillow~=10.0.0